[core] Cleanup nits from the #127 review
- discord_events.push_community_event: drop the redundant second
  SiteConfiguration.load() — from_settings() already folds
  discord_events_sync_enabled into client.enabled (one config query per event).
- Promote the shared Discord API base URL from private _API_BASE to public
  API_BASE in core.events.discord_dm; update every importer (no more
  underscore-prefixed cross-module import).
- seed_guild_calendar_events: preload existing import UIDs into a set in one
  query instead of a per-VEVENT exists() (adds newly created UIDs so a repeat
  within the file still skips).
- calendar_seed: drop the dead "guild meeting" keyword — "meeting" already
  matches it as a substring.
- Consolidate the run_scheduled_tasks dispatcher specs into
  tests/core/management/ (merging the unique gating / run-recording / failure
  coverage) and delete the stale core/spec/management/ copy.
